requirements.txt
Quart
yt-dlp
python-dotenv
gunicorn
asyncpg
psycopg
apscheduler